+++ mapie/tests/test_classification.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from typing import Any, Dict, Iterable, List, Optional, Union, cast
+from copy import deepcopy
 
 import numpy as np
 import pandas as pd
@@ -11,7 +12,7 @@ from sklearn.datasets import make_classification
 from sklearn.dummy import DummyClassifier
 from sklearn.impute import SimpleImputer
 from sklearn.linear_model import LogisticRegression
-from sklearn.model_selection import KFold, LeaveOneOut
+from sklearn.model_selection import KFold, LeaveOneOut, ShuffleSplit
 from sklearn.pipeline import Pipeline, make_pipeline
 from sklearn.preprocessing import OneHotEncoder
 from sklearn.utils.estimator_checks import check_estimator
@@ -76,6 +77,7 @@ Params = TypedDict(
     {
         "method": str,
         "cv": Optional[Union[int, str]],
+        "test_size": Optional[Union[int, float]],
         "random_state": Optional[int]
     }
 )
@@ -92,7 +94,20 @@ STRATEGIES = {
         Params(
             method="score",
             cv="prefit",
-            random_state=None
+            test_size=None,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=False,
+            agg_scores="mean"
+        )
+    ),
+    "score_split": (
+        Params(
+            method="score",
+            cv="split",
+            test_size=0.5,
+            random_state=42
         ),
         ParamsPredict(
             include_last_label=False,
@@ -103,7 +118,8 @@ STRATEGIES = {
         Params(
             method="score",
             cv=3,
-            random_state=None
+            test_size=None,
+            random_state=42
         ),
         ParamsPredict(
             include_last_label=False,
@@ -114,7 +130,8 @@ STRATEGIES = {
         Params(
             method="score",
             cv=3,
-            random_state=None
+            test_size=None,
+            random_state=42
         ),
         ParamsPredict(
             include_last_label=False,
@@ -125,6 +142,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv="prefit",
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -136,6 +154,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv="prefit",
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -147,6 +166,43 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv="prefit",
+            test_size=None,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label="randomized",
+            agg_scores="mean"
+        )
+    ),
+    "cumulated_score_include_split": (
+        Params(
+            method="cumulated_score",
+            cv="split",
+            test_size=0.5,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=True,
+            agg_scores="mean"
+        )
+    ),
+    "cumulated_score_not_include_split": (
+        Params(
+            method="cumulated_score",
+            cv="split",
+            test_size=0.5,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=False,
+            agg_scores="mean"
+        )
+    ),
+    "cumulated_score_randomized_split": (
+        Params(
+            method="cumulated_score",
+            cv="split",
+            test_size=0.5,
             random_state=42
         ),
         ParamsPredict(
@@ -158,6 +214,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv=3,
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -169,6 +226,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv=3,
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -180,6 +238,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv=3,
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -191,6 +250,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv=3,
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -202,6 +262,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv=3,
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -213,6 +274,7 @@ STRATEGIES = {
         Params(
             method="cumulated_score",
             cv=3,
+            test_size=None,
             random_state=42
         ),
         ParamsPredict(
@@ -224,6 +286,19 @@ STRATEGIES = {
         Params(
             method="naive",
             cv="prefit",
+            test_size=None,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=True,
+            agg_scores="mean"
+        )
+    ),
+    "naive_split": (
+        Params(
+            method="naive",
+            cv="split",
+            test_size=0.5,
             random_state=42
         ),
         ParamsPredict(
@@ -235,6 +310,19 @@ STRATEGIES = {
         Params(
             method="top_k",
             cv="prefit",
+            test_size=None,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=True,
+            agg_scores="mean"
+        )
+    ),
+    "top_k_split": (
+        Params(
+            method="top_k",
+            cv="split",
+            test_size=0.5,
             random_state=42
         ),
         ParamsPredict(
@@ -246,6 +334,19 @@ STRATEGIES = {
         Params(
             method="raps",
             cv="prefit",
+            test_size=None,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label=True,
+            agg_scores="mean"
+        )
+    ),
+    "raps_split": (
+        Params(
+            method="raps",
+            cv="split",
+            test_size=0.5,
             random_state=42
         ),
         ParamsPredict(
@@ -257,6 +358,19 @@ STRATEGIES = {
         Params(
             method="raps",
             cv="prefit",
+            test_size=None,
+            random_state=42
+        ),
+        ParamsPredict(
+            include_last_label="randomized",
+            agg_scores="mean"
+        )
+    ),
+    "raps_randomized_split": (
+        Params(
+            method="raps",
+            cv="split",
+            test_size=0.5,
             random_state=42
         ),
         ParamsPredict(
@@ -267,26 +381,35 @@ STRATEGIES = {
 }
 
 COVERAGES = {
-    "score": 6 / 9,
-    "score_cv_mean": 1,
-    "score_cv_crossval": 1,
-    "cumulated_score_include": 1,
-    "cumulated_score_not_include": 5 / 9,
-    "cumulated_score_randomized": 6 / 9,
-    "cumulated_score_include_cv_mean": 1,
-    "cumulated_score_not_include_cv_mean": 5 / 9,
-    "cumulated_score_randomized_cv_mean": 5 / 9,
-    "cumulated_score_include_cv_crossval": 2 / 9,
-    "cumulated_score_not_include_cv_crossval": 0,
-    "cumulated_score_randomized_cv_crossval": 4 / 9,
-    "naive": 5 / 9,
-    "top_k": 1,
-    "raps": 1,
-    "raps_randomized": 8/9
+    "score": 6/9,
+    "score_split": 8/9,
+    "score_cv_mean": 1.0,
+    "score_cv_crossval": 1.0,
+    "cumulated_score_include": 1.0,
+    "cumulated_score_not_include": 5/9,
+    "cumulated_score_randomized": 6/9,
+    "cumulated_score_include_split": 8/9,
+    "cumulated_score_not_include_split": 5/9,
+    "cumulated_score_randomized_split": 7/9,
+    "cumulated_score_include_cv_mean": 1.0,
+    "cumulated_score_not_include_cv_mean": 5/9,
+    "cumulated_score_randomized_cv_mean": 8/9,
+    "cumulated_score_include_cv_crossval": 4/9,
+    "cumulated_score_not_include_cv_crossval": 1/9,
+    "cumulated_score_randomized_cv_crossval": 7/9,
+    "naive": 5/9,
+    "naive_split": 5/9,
+    "top_k": 1.0,
+    "top_k_split": 1.0,
+    "raps": 1.0,
+    "raps_split": 7/9,
+    "raps_randomized": 8/9,
+    "raps_randomized_split": 1.0
 }
 
 X_toy = np.arange(9).reshape(-1, 1)
 y_toy = np.array([0, 0, 1, 0, 1, 1, 2, 1, 2])
+y_toy_string = np.array(["0", "0", "1", "0", "1", "1", "2", "1", "2"])
 
 y_toy_mapie = {
     "score": [
@@ -298,6 +421,17 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, False],
         [False, True, True],
+        [False, False, True]
+    ],
+    "score_split": [
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [False, True, True],
+        [False, False, True],
+        [False, False, True],
         [False, False, True],
     ],
     "score_cv_mean": [
@@ -306,10 +440,10 @@ y_toy_mapie = {
         [True, True, False],
         [True, True, False],
         [False, True, False],
-        [False, True, False],
         [False, True, True],
         [False, True, True],
         [False, True, True],
+        [False, True, True]
     ],
     "score_cv_crossval": [
         [True, False, False],
@@ -320,7 +454,7 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, True],
         [False, True, True],
-        [False, True, True],
+        [False, True, True]
     ],
     "cumulated_score_include": [
         [True, False, False],
@@ -331,7 +465,7 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, True],
         [False, True, True],
-        [False, False, True],
+        [False, False, True]
     ],
     "cumulated_score_not_include": [
         [True, False, False],
@@ -342,7 +476,7 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, False],
         [False, False, True],
-        [False, False, True],
+        [False, False, True]
     ],
     "cumulated_score_randomized": [
         [True, False, False],
@@ -353,18 +487,51 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, False],
         [False, True, True],
+        [False, False, True]
+    ],
+    "cumulated_score_include_split": [
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, True],
+        [True, True, True],
+        [False, True, True],
         [False, False, True],
+        [False, False, True]
+    ],
+    "cumulated_score_not_include_split": [
+        [False, True, False],
+        [False, True, False],
+        [False, True, False],
+        [False, True, False],
+        [True, True, False],
+        [False, True, True],
+        [False, False, True],
+        [False, False, True],
+        [False, False, True]
+    ],
+    "cumulated_score_randomized_split": [
+        [False, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [False, True, True],
+        [False, False, True],
+        [False, False, True],
+        [False, False, True]
     ],
     "cumulated_score_include_cv_mean": [
         [True, False, False],
-        [True, False, False],
         [True, True, False],
         [True, True, False],
         [True, True, False],
-        [False, True, False],
+        [True, True, False],
         [False, True, True],
         [False, True, True],
         [False, True, True],
+        [False, True, True]
     ],
     "cumulated_score_not_include_cv_mean": [
         [True, False, False],
@@ -374,52 +541,52 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, False],
         [False, True, False],
-        [False, True, False],
-        [False, True, False],
+        [False, False, True],
+        [False, False, True]
     ],
     "cumulated_score_randomized_cv_mean": [
         [True, False, False],
-        [True, False, False],
-        [True, False, False],
-        [False, True, False],
-        [False, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
         [False, True, False],
         [False, True, False],
         [False, True, True],
-        [False, True, False],
+        [False, True, True]
     ],
     "cumulated_score_include_cv_crossval": [
-        [False, False, False],
         [False, False, False],
         [True, False, False],
         [False, False, False],
         [False, True, False],
+        [True, True, False],
         [False, True, False],
         [False, True, False],
-        [False, False, False],
-        [False, False, False],
+        [False, True, False],
+        [False, False, False]
     ],
     "cumulated_score_not_include_cv_crossval": [
         [False, False, False],
         [False, False, False],
         [False, False, False],
         [False, False, False],
+        [False, True, False],
         [False, False, False],
         [False, False, False],
         [False, False, False],
-        [False, False, False],
-        [False, False, False],
+        [False, False, False]
     ],
     "cumulated_score_randomized_cv_crossval": [
         [True, False, False],
-        [False, False, False],
+        [True, False, False],
         [True, False, False],
         [False, True, False],
-        [False, True, False],
-        [False, True, False],
+        [True, True, True],
         [False, True, True],
-        [False, False, True],
+        [False, True, True],
         [False, True, False],
+        [False, False, True]
     ],
     "naive": [
         [True, False, False],
@@ -430,7 +597,18 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, False],
         [False, False, True],
+        [False, False, True]
+    ],
+    "naive_split": [
+        [False, True, False],
+        [False, True, False],
+        [False, True, False],
+        [False, True, False],
+        [False, True, False],
+        [False, True, True],
         [False, False, True],
+        [False, False, True],
+        [False, False, True]
     ],
     "top_k": [
         [True, True, False],
@@ -441,7 +619,18 @@ y_toy_mapie = {
         [False, True, True],
         [False, True, True],
         [False, True, True],
+        [False, True, True]
+    ],
+    "top_k_split": [
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
         [False, True, True],
+        [False, True, True],
+        [False, True, True],
+        [False, True, True]
     ],
     "raps": [
         [True, False, False],
@@ -452,7 +641,18 @@ y_toy_mapie = {
         [False, True, True],
         [False, True, True],
         [False, True, True],
-        [False, True, True],
+        [False, True, True]
+    ],
+    "raps_split": [
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False],
+        [True, True, False]
     ],
     "raps_randomized": [
         [True, False, False],
@@ -463,8 +663,19 @@ y_toy_mapie = {
         [False, True, False],
         [False, True, False],
         [False, True, True],
-        [False, False, True],
+        [False, False, True]
     ],
+    "raps_randomized_split": [
+        [True, True, True],
+        [True, True, True],
+        [True, True, True],
+        [True, True, True],
+        [True, True, True],
+        [True, True, True],
+        [True, True, True],
+        [True, True, True],
+        [True, True, True]
+    ]
 }
 
 REGULARIZATION_PARAMETERS = [
@@ -633,9 +844,10 @@ def test_default_parameters() -> None:
     assert mapie_clf.method == "score"
 
 
-def test_warning_binary_classif() -> None:
+@pytest.mark.parametrize("cv", ["prefit", "split"])
+def test_warning_binary_classif(cv: str) -> None:
     """Test that a warning is raised y is binary."""
-    mapie_clf = MapieClassifier(random_state=42)
+    mapie_clf = MapieClassifier(cv=cv, random_state=42)
     X, y = make_classification(
         n_samples=500,
         n_features=10,
@@ -680,7 +892,10 @@ def test_valid_method(method: str) -> None:
     check_is_fitted(mapie_clf, mapie_clf.fit_attributes)
 
 
-@pytest.mark.parametrize("cv", [None, -1, 2, KFold(), LeaveOneOut()])
+@pytest.mark.parametrize(
+    "cv", [None, -1, 2, KFold(), LeaveOneOut(),
+           ShuffleSplit(n_splits=1, test_size=0.5), "prefit"]
+)
 def test_valid_cv(cv: Any) -> None:
     """Test that valid cv raises no errors."""
     model = LogisticRegression(multi_class="multinomial")
@@ -756,6 +971,138 @@ def test_predict_output_shape(
     n_alpha = len(alpha) if hasattr(alpha, "__len__") else 1
     assert y_pred.shape == (X.shape[0],)
     assert y_ps.shape == (X.shape[0], len(np.unique(y)), n_alpha)
+
+
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+@pytest.mark.parametrize("alpha", [0.2, [0.2, 0.3], (0.2, 0.3)])
+def test_y_is_list_of_string(
+    strategy: str, alpha: Any,
+) -> None:
+    """Test predict output shape with string y."""
+    args_init, args_predict = STRATEGIES[strategy]
+    mapie_clf = MapieClassifier(**args_init)
+    mapie_clf.fit(X, y.astype('str'))
+    y_pred, y_ps = mapie_clf.predict(
+        X,
+        alpha=alpha,
+        include_last_label=args_predict["include_last_label"],
+        agg_scores=args_predict["agg_scores"]
+    )
+    n_alpha = len(alpha) if hasattr(alpha, "__len__") else 1
+    assert y_pred.shape == (X.shape[0],)
+    assert y_ps.shape == (X.shape[0], len(np.unique(y)), n_alpha)
+
+
+@pytest.mark.parametrize(
+    "strategy", ["naive", "top_k", "score", "cumulated_score_include"]
+)
+def test_same_results_prefit_split(strategy: str) -> None:
+    """
+    Test checking that if split and prefit method have exactly
+    the same data split, then we have exactly the same results.
+    """
+    random_state = 1
+    X, y = make_classification(
+        n_samples=500,
+        n_features=10,
+        n_informative=3,
+        n_classes=n_classes,
+        random_state=random_state,
+    )
+    cv = ShuffleSplit(n_splits=1, test_size=0.5, random_state=random_state)
+    train_index, val_index = next(cv.split(X))
+    X_train_, X_calib_ = X[train_index], X[val_index]
+    y_train_, y_calib_ = y[train_index], y[val_index]
+
+    args_init, args_predict = deepcopy(STRATEGIES[strategy + '_split'])
+    args_init["cv"] = cv
+    mapie_reg = MapieClassifier(**args_init)
+    mapie_reg.fit(X, y)
+    y_pred_1, y_pis_1 = mapie_reg.predict(X, alpha=0.1, **args_predict)
+
+    args_init, _ = STRATEGIES[strategy]
+    model = LogisticRegression().fit(X_train_, y_train_)
+    mapie_reg = MapieClassifier(estimator=model, **args_init)
+    mapie_reg.fit(X_calib_, y_calib_)
+    y_pred_2, y_pis_2 = mapie_reg.predict(X, alpha=0.1, **args_predict)
+
+    np.testing.assert_allclose(y_pred_1, y_pred_2)
+    np.testing.assert_allclose(y_pis_1[:, 0, 0], y_pis_2[:, 0, 0])
+    np.testing.assert_allclose(y_pis_1[:, 1, 0], y_pis_2[:, 1, 0])
+
+
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+@pytest.mark.parametrize("alpha", [0.2, [0.2, 0.3], (0.2, 0.3)])
+def test_same_result_y_numeric_and_string(
+    strategy: str, alpha: Any,
+) -> None:
+    """Test that MAPIE outputs the same results if y is
+    numeric or string"""
+    args_init, args_predict = STRATEGIES[strategy]
+    mapie_clf_str = MapieClassifier(**args_init)
+    mapie_clf_str.fit(X, y.astype('str'))
+    mapie_clf_int = MapieClassifier(**args_init)
+    mapie_clf_int.fit(X, y)
+    _, y_ps_str = mapie_clf_str.predict(
+        X,
+        alpha=alpha,
+        include_last_label=args_predict["include_last_label"],
+        agg_scores=args_predict["agg_scores"],
+    )
+    _, y_ps_int = mapie_clf_int.predict(
+        X,
+        alpha=alpha,
+        include_last_label=args_predict["include_last_label"],
+        agg_scores=args_predict["agg_scores"]
+    )
+    np.testing.assert_allclose(y_ps_int, y_ps_str)
+
+
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+@pytest.mark.parametrize("alpha", [0.2, [0.2, 0.3], (0.2, 0.3)])
+def test_y_1_to_l_minus_1(
+    strategy: str, alpha: Any,
+) -> None:
+    """Test predict output shape with string y."""
+    args_init, args_predict = STRATEGIES[strategy]
+    mapie_clf = MapieClassifier(**args_init)
+    mapie_clf.fit(X, y + 1)
+    y_pred, y_ps = mapie_clf.predict(
+        X,
+        alpha=alpha,
+        include_last_label=args_predict["include_last_label"],
+        agg_scores=args_predict["agg_scores"]
+    )
+    n_alpha = len(alpha) if hasattr(alpha, "__len__") else 1
+    assert y_pred.shape == (X.shape[0],)
+    assert y_ps.shape == (X.shape[0], len(np.unique(y)), n_alpha)
+
+
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+@pytest.mark.parametrize("alpha", [0.2, [0.2, 0.3], (0.2, 0.3)])
+def test_same_result_y_numeric_and_1_to_l_minus_1(
+    strategy: str, alpha: Any,
+) -> None:
+    """Test that MAPIE outputs the same results if y is
+    numeric or string"""
+    args_init, args_predict = STRATEGIES[strategy]
+    mapie_clf_1 = MapieClassifier(**args_init)
+    mapie_clf_1.fit(X, y + 1)
+    mapie_clf_int = MapieClassifier(**args_init)
+    mapie_clf_int.fit(X, y)
+    _, y_ps_1 = mapie_clf_1.predict(
+        X,
+        alpha=alpha,
+        include_last_label=args_predict["include_last_label"],
+        agg_scores=args_predict["agg_scores"],
+    )
+    _, y_ps_int = mapie_clf_int.predict(
+        X,
+        alpha=alpha,
+        include_last_label=args_predict["include_last_label"],
+        agg_scores=args_predict["agg_scores"]
+    )
+    np.testing.assert_allclose(y_ps_int, y_ps_1)
 
 
 @pytest.mark.parametrize("strategy", [*STRATEGIES])
@@ -897,7 +1244,10 @@ def test_valid_prediction(alpha: Any) -> None:
 def test_toy_dataset_predictions(strategy: str) -> None:
     """Test prediction sets estimated by MapieClassifier on a toy dataset"""
     args_init, args_predict = STRATEGIES[strategy]
-    clf = LogisticRegression().fit(X_toy, y_toy)
+    if "split" not in strategy:
+        clf = LogisticRegression().fit(X_toy, y_toy)
+    else:
+        clf = LogisticRegression()
     mapie_clf = MapieClassifier(estimator=clf, **args_init)
     mapie_clf.fit(X_toy, y_toy, size_raps=.5)
     _, y_ps = mapie_clf.predict(
@@ -1270,7 +1620,7 @@ def test_get_true_label_position(y_true_proba_place: List[NDArray]):
 @pytest.mark.parametrize("cv", [5, None])
 def test_error_raps_cv_not_prefit(cv: Union[int, None]):
     """Test that an error is raised if the method is RAPS
-    and cv is different from prefit.
+    and cv is different from prefit and split.
     """
     mapie = MapieClassifier(method="raps", cv=5)
     with pytest.raises(ValueError, match=r".*RAPS method can only.*"):